Fix bug in makefile/cmake symboli clink handling
The problem is, that gnu make resolves $curdir always to the true absolute path.
If the build directory is a symbolic link, the path is resolved before starting
main. However, CMake is invoked in a new shell which inherits CWD (the symbolic
link). This causes 1000s of cmake errors because the GVSOC_MODULES path is not
the same folder, cmake tries to build it in. By enforcing the wrapper makefile
to use a single shell and changing the CWD to the resolved absolute path, GNU
Make and CMake use the same path for cwd (the resolved one) and the error does
no longer occur.