Add basic CI flow (#1)
* WIP add github actions CI

* Update instructions in readme to initialize the submodules first

* Add CI check for ETH servers

* Fix typo in README

* Fix syntax errors in workflow

* Use python3.9 for ETH CI

* Fix bug in makefile/cmake symboli clink handling

The problem is, that gnu make resolves $curdir always to the true absolute path.
If the build directory is a symbolic link, the path is resolved before starting
main. However, CMake is invoked in a new shell which inherits CWD (the symbolic
link). This causes 1000s of cmake errors because the GVSOC_MODULES path is not
the same folder, cmake tries to build it in. By enforcing the wrapper makefile
to use a single shell and changing the CWD to the resolved absolute path, GNU
Make and CMake use the same path for cwd (the resolved one) and the error does
no longer occur.

* Fixed wrong test gapy command

* Fixed rv64

* Fixed wrong target for build in ci

* Fixed wrong run command in gitlab ci

* Removed .ONESHELL to still stop on command erorr

---------

Co-authored-by: Germain Haugou <germain.haugou@greenwaves-technologies.com>